Refactored settings
