+++ app/src/main/res/menu/main_activity.xml
@@ -13,8 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 -->
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
         android:id="@+id/settings"
-        android:title="@string/settings"/>
+        android:title="@string/settings"
+        android:icon="@drawable/ic_settings"
+        app:showAsAction="always" />
 </menu>
